+++ R/util.R
@@ -884,3 +884,124 @@ process_factor_unit <- function(vertex_unit, length_unit){
     return(fact)
   }
 }
+
+
+#' @noRd 
+#' code from https://gist.github.com/MansMeg/1ec56b54e1d9d238b4fd
+#' 
+#' Message progress bar
+#' 
+#' @description 
+#' A simple progress bar to use in R packages where messages are prefered to console output.
+#' 
+#' @field iter Total number of iterations
+#' @field i Current iteration
+#' @field width Width of the R console
+#' @field width_bar Width of the progress bar
+#' @field progress The number of character printed (continous)
+#' @field progress_step Addition to progress per iteration
+#' 
+#' @examples
+#' test_bar <- function(i = 10){
+#'  bar <- msg_progress_bar(i)
+#'  for(j in 1:i){
+#'    bar$increment()
+#'    Sys.sleep(4/i)
+#'    }
+#'  }
+#'  test_bar(100)
+#'   
+#' @author Mans Magnusson (MansMeg @ github)
+#'   
+msg_progress_bar <- 
+  setRefClass(
+    Class = "msg_progress_bar", 
+    fields = list(iter = "numeric",
+                  i = "numeric",
+                  progress = "numeric",
+                  progress_step = "numeric",
+                  width = "numeric",
+                  width_bar = "numeric"),
+    
+    methods = list(
+      initialize = function(iter){
+        'Initialize a messagebar object'
+        .self$width <- getOption("width")
+        .self$iter <- iter
+        .self$i <- 0
+        .self$progress <- 0
+        white_part <- paste(rep(" ", (.self$width - 11) %/% 4), collapse="")
+        init_length <- .self$width - ((.self$width - 11) %/% 4) * 4 - 11
+        white_init <- paste(rep(" ", init_length), collapse="")
+        .self$width_bar <- .self$width - init_length - 2 + 0.1
+        .self$progress_step <- .self$width_bar / .self$iter
+        message(paste(white_init, "|", white_part, "25%", white_part, "50%", white_part, "75%", white_part, "|","\n", white_init, "|", sep=""), appendLF = FALSE)
+      },
+      
+      increment = function(){
+        'A messagebar object.'
+        if(.self$i > .self$iter) return(invisible(NULL))
+        new_progress <- .self$progress + .self$progress_step
+        diff_in_char <- floor(new_progress) - floor(.self$progress)
+        if(diff_in_char > 0) {
+          message(paste(rep("=", diff_in_char),collapse=""), appendLF = FALSE)
+        }
+        
+        .self$progress <- new_progress
+        .self$i <- .self$i + 1
+        if(.self$i == .self$iter) message("|\n", appendLF = FALSE)
+        
+      }
+    )
+  )
+
+  #' @noRd
+  #' 
+  
+  get_rel_pos_prune <- function(which_line_starts, Line_1, Line_2, start_1, end_1, start_2, end_2, length_line1, length_line2){
+
+      if(Line_1 != Line_2){
+        total_length <- length_line2 + length_line1
+        if(which_line_starts == 1){
+          # if(start_2 == 0) {
+          #   end <- (end_2*length_line2+length_line1)/total_length
+          # } else {
+          #   end <- ((end_2 - start_2)*length_line2+length_line1)/total_length
+          # }
+          # if(end_1 == 1){
+          #   start <- (start_1 * length_line1)/total_length
+          # } else{
+          #   start <- ((end_1 - start_1)*length_line1+length_line2)/total_length
+          # }
+          start <- start_1 * length_line1 / total_length
+          end <- (end_2 * length_line2 + length_line1)/ total_length
+        } else{
+          # if(start_1 == 0) {
+          #   end <- (end_1*length_line1+length_line2)/total_length
+          # } else {
+          #   end <- ((end_1 - start_1)*length_line1+length_line2)/total_length
+          # }
+          # if(end_2 == 1){
+          #   start <- (start_2 * length_line2)/total_length
+          # } else{
+          #   start <- ((end_2 - start_2)*length_line2+length_line1)/total_length
+          # }
+          start <- start_2 * length_line2 / total_length
+          end <- (end_1 * length_line1 + length_line2)/ total_length
+        }
+      } else {
+          end <- max(end_1,end_2)
+          start <- min(start_1,start_2)
+      }
+
+      return(list(start = start, end = end))
+  }
+
+
+
+#' @noRd 
+#' 
+
+get_vertex_pos_in_line <- function(V, coords_line){
+    return(which.min(sapply(1:nrow(coords_line), function(i){norm(as.matrix(V - coords_line[i,]))})))
+}